chat-bot/requirement.txt
# LangChain and Ecosystem
langchain == 0.3.25
langchain-community == 0.3.24
langchain-huggingface == 0.2.0
openai== 1.78.1
tiktoken == 0.9.0
chromadb == 1.0.9

# PDF and Document Processing
PyMuPDF==1.25.5
pdfplumber==0.11.6
PyPDF2==3.0.1
unstructured-inference== 0.8.10

# OCR
pytesseract==0.3.13

# Embeddings and Vector Search
sentence-transformers==  4.1.0
faiss-cpu==1.11.0

# General ML / NLP
transformers
scikit-learn==1.4.2
scipy== 1.15.3

# Basic Data Libraries
pandas==2.2.3
numpy==2.2.6
matplotlib==3.10.3
requests==2.32.3
python-dotenv==1.1.0
spacy==3.8.7
datasets==3.6.0